Refactor line item handling towards unifying interface
