+++ src/Payment/LineItems/PaymentLines.php
@@ -2,16 +2,17 @@
 
 declare(strict_types=1);
 
-namespace Mollie\WooCommerce\Payment;
+namespace Mollie\WooCommerce\Payment\LineItems;
 
 use Mollie\WooCommerce\PaymentMethods\Constants;
 use Mollie\WooCommerce\PaymentMethods\Voucher;
 use Mollie\WooCommerce\Shared\Data;
 use WC_Order;
 use WC_Order_Item;
+use WC_Order_Item_Fee;
 use WC_Tax;
 
-class OrderLines
+class PaymentLines implements LineItemProvider
 {
     /**
      * Formatted order lines.
@@ -56,7 +57,7 @@ class OrderLines
      *
      * @return array
      */
-    public function order_lines($order)
+    public function order_lines(WC_Order $order): array
     {
         $this->order_lines = [];
         $this->order = $order;
@@ -67,9 +68,7 @@ class OrderLines
         $this->process_gift_cards();
         $this->process_mismatch();
 
-        return  [
-            'lines' => $this->get_order_lines(),
-        ];
+        return $this->get_order_lines();
     }
 
     private function process_mismatch()
@@ -86,7 +85,7 @@ class OrderLines
         }
         $mismatch =  [
             'type' => $orderTotalDiff > 0 ? 'surcharge' : 'discount',
-            'name' => __('Rounding difference', 'mollie-payments-for-woocommerce'),
+            'description' => __('Rounding difference', 'mollie-payments-for-woocommerce'),
             'quantity' => 1,
             'vatRate' => 0,
             'unitPrice' =>  [
@@ -101,11 +100,7 @@ class OrderLines
                 'currency' => $this->currency,
                 'value' => $this->dataHelper->formatCurrencyValue(0, $this->currency),
             ],
-            'metadata' =>  [
-                'order_item_id' => 'rounding_diff',
-            ],
         ];
-
         $this->order_lines[] = $mismatch;
     }
 
@@ -138,41 +133,35 @@ class OrderLines
                 }
 
                 $this->currency = $this->dataHelper->getOrderCurrency($this->order);
-                $vatRate = round($this->get_item_vatRate($cart_item, $product), 2);
-                $wcTotalValue = $this->get_item_total_amount($cart_item);
-                $wcUnitPrice = $this->get_item_price($cart_item);
 
-                // Calculate Mollie prices, they expect price including VAT
-                $mollieTotal = $this->getMolliePrice($wcTotalValue, $vatRate);
-                $mollieUnit = $this->getMolliePrice($wcUnitPrice, $vatRate);
+                $vatRate = 0;
+                if ($cart_item['line_tax'] > 0 && $cart_item['line_total'] > 0) {
+                    $vatRate = round($cart_item['line_tax'] / $cart_item['line_total'], 4) * 100;
+                }
 
                 $mollie_order_item =  [
                     'sku' => $this->get_item_reference($product),
                     'type' => ($product instanceof \WC_Product && $product->is_virtual()) ? 'digital' : 'physical',
-                    'name' => $this->get_item_name($cart_item),
+                    'description' => $this->get_item_name($cart_item),
                     'quantity' => $this->get_item_quantity($cart_item),
                     'vatRate' => $vatRate,
                     'unitPrice' =>  [
                         'currency' => $this->currency,
-                        'value' => $this->dataHelper->formatCurrencyValue($mollieUnit['grossPrice'], $this->currency),
+                        'value' => $this->dataHelper->formatCurrencyValue($this->get_item_price($cart_item), $this->currency),
                     ],
                     'totalAmount' =>  [
                         'currency' => $this->currency,
-                        'value' => $this->dataHelper->formatCurrencyValue($mollieTotal['grossPrice'], $this->currency),
+                        'value' => $this->dataHelper->formatCurrencyValue($this->get_item_total_amount($cart_item), $this->currency),
                     ],
                     'vatAmount' =>
                          [
                             'currency' => $this->currency,
-                            'value' => $this->dataHelper->formatCurrencyValue($mollieTotal['vatAmount'], $this->currency),
+                            'value' => $this->dataHelper->formatCurrencyValue($this->get_item_tax_amount($cart_item), $this->currency),
                         ],
                     'discountAmount' =>
                          [
                             'currency' => $this->currency,
                             'value' => $this->dataHelper->formatCurrencyValue($this->get_item_discount_amount($cart_item), $this->currency),
-                        ],
-                    'metadata' =>
-                        [
-                            'order_item_id' => $cart_item->get_id(),
                         ],
                     'productUrl' => ($product instanceof \WC_Product) ? $product->get_permalink() : null,
                 ];
@@ -194,7 +183,7 @@ class OrderLines
                 if ($paymentMethod === 'mollie_wc_gateway_' . Constants::VOUCHER && $product instanceof \WC_Product) {
                     $categories = Voucher::getCategoriesForProduct($product);
                     if ($categories) {
-                        $mollie_order_item['category'] = array_shift($categories);
+                        $mollie_order_item['categories'] = $categories;
                     }
                 }
                 $this->order_lines[] = $mollie_order_item;
@@ -220,7 +209,7 @@ class OrderLines
                 }
                 $shipping = [
                     'type' => 'shipping_fee',
-                    'name' => $shipping_method->get_name() ?: __('Shipping', 'mollie-payments-for-woocommerce'),
+                    'description' => $shipping_method->get_name() ?: __('Shipping', 'mollie-payments-for-woocommerce'),
                     'quantity' => 1,
                     'vatRate' => $vatRate,
                     'unitPrice' =>  [
@@ -234,9 +223,6 @@ class OrderLines
                     'vatAmount' =>  [
                         'currency' => $this->currency,
                         'value' => $this->dataHelper->formatCurrencyValue($shipping_method->get_total_tax(), $this->currency),
-                    ],
-                    'metadata' => [
-                        'order_item_id' => $shipping_method->get_id(),
                     ],
                 ];
 
@@ -254,6 +240,7 @@ class OrderLines
     {
         if (! empty($this->order->get_items('fee'))) {
             foreach ($this->order->get_items('fee') as $cart_fee) {
+                assert($cart_fee instanceof WC_Order_Item_Fee);
                 if ($cart_fee['tax_status'] === 'taxable') {
                     // Calculate tax rate.
                     $tmp_rates = WC_Tax::get_rates($cart_fee['tax_class']);
@@ -282,7 +269,7 @@ class OrderLines
 
                 $fee =  [
                     'type' => $cart_fee_total > 0 ? 'surcharge' : 'discount',
-                    'name' => $cart_fee['name'],
+                    'description' => $cart_fee['name'],
                     'quantity' => 1,
                     'vatRate' => $this->dataHelper->formatCurrencyValue($cart_fee_vat_rate, $this->currency),
                     'unitPrice' =>  [
@@ -296,9 +283,6 @@ class OrderLines
                     'vatAmount' =>  [
                         'currency' => $this->currency,
                         'value' => $this->dataHelper->formatCurrencyValue($cart_fee_tax_amount, $this->currency),
-                    ],
-                    'metadata' =>  [
-                        'order_item_id' => $cart_fee->get_id(),
                     ],
                 ];
 
@@ -318,7 +302,7 @@ class OrderLines
             foreach ($this->order->get_items('gift_card') as $cart_gift_card) {
                 $gift_card = [
                     'type' => 'gift_card',
-                    'name' => $cart_gift_card->get_name(),
+                    'description' => $cart_gift_card->get_name(),
                     'unitPrice' => [
                         'currency' => $this->currency,
                         'value' =>  $this->dataHelper->formatCurrencyValue(-$cart_gift_card->get_amount(), $this->currency),
@@ -371,41 +355,6 @@ class OrderLines
     private function get_item_tax_amount($cart_item)
     {
         return $cart_item['line_tax'];
-    }
-
-    /**
-     * Calculate item tax percentage.
-     *
-     * @since  1.0
-     * @access private
-     *
-     * @param  WC_Order_Item  $cart_item Cart item.
-     * @param  null|false|\WC_Product $product   Product object.
-     *
-     * @return integer $item_vatRate Item tax percentage formatted for Mollie Orders API.
-     */
-    private function get_item_vatRate($cart_item, $product)
-    {
-        if ($product && $product->is_taxable() && $cart_item['line_subtotal_tax'] > 0) {
-            // Calculate tax rate.
-            $_tax = new WC_Tax();
-            $tmp_rates = $_tax->get_rates($product->get_tax_class());
-            $item_vatRate = 0;
-            foreach ($tmp_rates as $rate) {
-                if (isset($rate['rate'])) {
-                    if ($rate['compound'] === "yes") {
-                        $compoundRate = round($item_vatRate * ($rate['rate'] / 100)) + $rate['rate'];
-                        $item_vatRate += $compoundRate;
-                        continue;
-                    }
-                    $item_vatRate += $rate['rate'];
-                }
-            }
-        } else {
-            $item_vatRate = 0;
-        }
-
-        return $item_vatRate;
     }
 
     /**
@@ -500,22 +449,5 @@ class OrderLines
     private function get_item_total_amount($cart_item)
     {
         return $cart_item['line_total'] + $cart_item['line_tax'];
-    }
-
-    /**
-     * Build price data for Mollie API.
-     *
-     * @param float $wcPrice
-     * @param float $vatRate
-     * @return float[]
-     */
-    private function getMolliePrice(float $wcPrice, float $vatRate): array
-    {
-        $grossPrice = wc_prices_include_tax() ? $wcPrice : $wcPrice * (1 + ($vatRate / 100));
-
-        return [
-            'grossPrice' => $grossPrice,
-            'vatAmount' => $grossPrice * ($vatRate / (100 + $vatRate)),
-        ];
     }
 }
